доделал скрипт: "копирование текущего файла", из hw05_easy.

+++ lesson_5/copy_current_file.py
@@ -1,14 +1,18 @@
 import os
 import shutil
 
+# Функция создаёт копию файла, из которого была запущена.
+
+
 def copy_file(filename):
-    # if os.path.isfile(filename):
-    newfile = filename + '.copy'
-    shutil.copy(filename, newfile)
-    if os.path.exists(newfile):
-        print("Файл ", newfile, " был успешно создан")
-    else:
-        print("Возникли проблемы копирования")
+    if os.path.isfile(filename):
+        newfile = filename + '.copy'
+        shutil.copy(filename, newfile)
+        if os.path.exists(newfile):
+            print(f"Файл {newfile} был успешно создан")
+        else:
+            print("Возникли проблемы при копировании")
 
 
-copy_file(os.path.abspath(__file__))
+if __name__ == '__main__':
+    copy_file(os.path.abspath(__file__))
